# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * School Timetable Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface LoginRequest {
  username: string;
  password: string;
}

export interface Admin {
  id: number;
  username: string;
  name: string;
  createdAt: string;
}

export interface LoginResponse {
  admin: Admin;
  message: string;
}

export interface School {
  id: number;
  name: string;
  address?: string;
  /** Type of school: primary, secondary, or custom */
  gradeType: string;
  minGrade: number;
  maxGrade: number;
  /** Comma-separated custom grade names */
  customGrades?: string;
  createdAt: string;
}

export interface CreateSchoolRequest {
  name: string;
  address?: string;
  gradeType: string;
  minGrade: number;
  maxGrade: number;
  customGrades?: string;
}

export interface Teacher {
  id: number;
  schoolId: number;
  name: string;
  email?: string;
  phone?: string;
  /** Comma-separated list of subjects */
  subjects?: string;
  createdAt: string;
}

export interface CreateTeacherRequest {
  name: string;
  email?: string;
  phone?: string;
  subjects?: string;
}

export interface Class {
  id: number;
  schoolId: number;
  grade: string;
  numSections: number;
  /** Comma-separated section names, e.g. A,B,C */
  sections: string;
  createdAt: string;
}

export interface CreateClassRequest {
  grade: string;
  numSections: number;
  sections: string;
}

export interface TimetableEntry {
  id: number;
  schoolId: number;
  classId: number;
  className?: string;
  teacherId: number;
  teacherName?: string;
  section: string;
  subject: string;
  /** Monday, Tuesday, Wednesday, Thursday, Friday, Saturday, Sunday */
  dayOfWeek: string;
  /** HH:MM format */
  startTime: string;
  /** HH:MM format */
  endTime: string;
  room?: string;
  createdAt: string;
}

export interface CreateTimetableEntryRequest {
  classId: number;
  teacherId: number;
  section: string;
  subject: string;
  dayOfWeek: string;
  startTime: string;
  endTime: string;
  room?: string;
}

export type ListTimetableEntriesParams = {
  classId?: number;
  section?: string;
};