# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: School Timetable Management API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: schools
    description: School management
  - name: teachers
    description: Teacher management
  - name: classes
    description: Class management
  - name: timetable
    description: Timetable management
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current admin
      responses:
        "200":
          description: Current admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Admin"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /schools:
    get:
      operationId: listSchools
      tags: [schools]
      summary: List all schools
      responses:
        "200":
          description: List of schools
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/School"
    post:
      operationId: createSchool
      tags: [schools]
      summary: Create a school
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSchoolRequest"
      responses:
        "201":
          description: School created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/School"

  /schools/{schoolId}:
    get:
      operationId: getSchool
      tags: [schools]
      summary: Get a school
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: School details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/School"
    put:
      operationId: updateSchool
      tags: [schools]
      summary: Update a school
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSchoolRequest"
      responses:
        "200":
          description: Updated school
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/School"
    delete:
      operationId: deleteSchool
      tags: [schools]
      summary: Delete a school
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /schools/{schoolId}/teachers:
    get:
      operationId: listTeachers
      tags: [teachers]
      summary: List teachers in a school
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of teachers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Teacher"
    post:
      operationId: createTeacher
      tags: [teachers]
      summary: Add a teacher to a school
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTeacherRequest"
      responses:
        "201":
          description: Teacher created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teacher"

  /schools/{schoolId}/teachers/{teacherId}:
    put:
      operationId: updateTeacher
      tags: [teachers]
      summary: Update a teacher
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
        - name: teacherId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTeacherRequest"
      responses:
        "200":
          description: Updated teacher
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teacher"
    delete:
      operationId: deleteTeacher
      tags: [teachers]
      summary: Delete a teacher
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
        - name: teacherId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /schools/{schoolId}/classes:
    get:
      operationId: listClasses
      tags: [classes]
      summary: List classes in a school
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of classes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Class"
    post:
      operationId: createClass
      tags: [classes]
      summary: Create a class in a school
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClassRequest"
      responses:
        "201":
          description: Class created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Class"

  /schools/{schoolId}/classes/{classId}:
    put:
      operationId: updateClass
      tags: [classes]
      summary: Update a class
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
        - name: classId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClassRequest"
      responses:
        "200":
          description: Updated class
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Class"
    delete:
      operationId: deleteClass
      tags: [classes]
      summary: Delete a class
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
        - name: classId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /schools/{schoolId}/timetable:
    get:
      operationId: listTimetableEntries
      tags: [timetable]
      summary: List timetable entries for a school
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
        - name: classId
          in: query
          schema:
            type: integer
        - name: section
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Timetable entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimetableEntry"
    post:
      operationId: createTimetableEntry
      tags: [timetable]
      summary: Create a timetable entry
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTimetableEntryRequest"
      responses:
        "201":
          description: Timetable entry created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimetableEntry"

  /schools/{schoolId}/timetable/{entryId}:
    put:
      operationId: updateTimetableEntry
      tags: [timetable]
      summary: Update a timetable entry
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
        - name: entryId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTimetableEntryRequest"
      responses:
        "200":
          description: Updated entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimetableEntry"
    delete:
      operationId: deleteTimetableEntry
      tags: [timetable]
      summary: Delete a timetable entry
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
        - name: entryId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success

    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password

    LoginResponse:
      type: object
      properties:
        admin:
          $ref: "#/components/schemas/Admin"
        message:
          type: string
      required:
        - admin
        - message

    Admin:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        name:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - username
        - name
        - createdAt

    School:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        address:
          type: string
        gradeType:
          type: string
          description: "Type of school: primary, secondary, or custom"
        minGrade:
          type: integer
        maxGrade:
          type: integer
        customGrades:
          type: string
          description: "Comma-separated custom grade names"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - gradeType
        - minGrade
        - maxGrade
        - createdAt

    CreateSchoolRequest:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        gradeType:
          type: string
        minGrade:
          type: integer
        maxGrade:
          type: integer
        customGrades:
          type: string
      required:
        - name
        - gradeType
        - minGrade
        - maxGrade

    Teacher:
      type: object
      properties:
        id:
          type: integer
        schoolId:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        subjects:
          type: string
          description: "Comma-separated list of subjects"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - schoolId
        - name
        - createdAt

    CreateTeacherRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        subjects:
          type: string
      required:
        - name

    Class:
      type: object
      properties:
        id:
          type: integer
        schoolId:
          type: integer
        grade:
          type: string
        numSections:
          type: integer
        sections:
          type: string
          description: "Comma-separated section names, e.g. A,B,C"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - schoolId
        - grade
        - numSections
        - sections
        - createdAt

    CreateClassRequest:
      type: object
      properties:
        grade:
          type: string
        numSections:
          type: integer
        sections:
          type: string
      required:
        - grade
        - numSections
        - sections

    TimetableEntry:
      type: object
      properties:
        id:
          type: integer
        schoolId:
          type: integer
        classId:
          type: integer
        className:
          type: string
        teacherId:
          type: integer
        teacherName:
          type: string
        section:
          type: string
        subject:
          type: string
        dayOfWeek:
          type: string
          description: "Monday, Tuesday, Wednesday, Thursday, Friday, Saturday, Sunday"
        startTime:
          type: string
          description: "HH:MM format"
        endTime:
          type: string
          description: "HH:MM format"
        room:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - schoolId
        - classId
        - teacherId
        - section
        - subject
        - dayOfWeek
        - startTime
        - endTime
        - createdAt

    CreateTimetableEntryRequest:
      type: object
      properties:
        classId:
          type: integer
        teacherId:
          type: integer
        section:
          type: string
        subject:
          type: string
        dayOfWeek:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        room:
          type: string
      required:
        - classId
        - teacherId
        - section
        - subject
        - dayOfWeek
        - startTime
        - endTime