# lib/api-zod/src/generated/types/admin.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * School Timetable Management API
 * OpenAPI spec version: 0.1.0
 */

export interface Admin {
  id: number;
  username: string;
  name: string;
  createdAt: Date;
}