# lib/api-zod/src/generated/types/class.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * School Timetable Management API
 * OpenAPI spec version: 0.1.0
 */

export interface Class {
  id: number;
  schoolId: number;
  grade: string;
  numSections: number;
  /** Comma-separated section names, e.g. A,B,C */
  sections: string;
  createdAt: Date;
}