# lib/api-zod/src/generated/types/createSchoolRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * School Timetable Management API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateSchoolRequest {
  name: string;
  address?: string;
  gradeType: string;
  minGrade: number;
  maxGrade: number;
  customGrades?: string;
}